dirty hack to fix Squiz username
The Squiz/Sugar ssh functions assume that the localhost username (on
your laptop) is the same as your general Squiz LDAP username. For many,
this is true, but for me, I've set my laptop username to be my initials.

So my dirty hack just has my Squiz user hard-coded in $squiz_user.  To
make substitution into the ssh commands simpler, the variable has the @
sign appended in it, so that if the variable is unset, it's not an
error.