Separate Cowboy's writings out from README.md
I want to start making these dotfiles my own. While I feel that it's important
to retain a link back to the source material, it doesn't really help me to have
it in the project README, so I've moved it to a new HISTORY.md and made the
README more specific (with reference back to Cowboy's repo for getting to the
true source, and in case anyone is thinking to fork this, which wouldn't be
helpful for them).