Update copyright notice
This is still Be Alman's work.  But Emacs keeps prompting to update the
Copyright notice date, and there have been some tweaks worth clarifying.

I consider CC-BY-4.0 to be functionally equivalent to MIT, but without the need
to write my own version of the MIT license with my name in it.