source vendor/z for fast directory switch, fix jobs count in prompt
z.sh looks like it would be more useful if it were actually sourced, so I
removed the comment (and I prefer 'source' over '.').  This exposed a bug in may
prompt function to count background jobs -- it would show (26) in source-control
working copies.

I traced this to how z.sh  works: it sets up some sort of background task to do
the "frecency" ranking, and this was being picked up in the call to the 'jobs'
builtin.

So I refined jobs to be jobs -r (just Running jobs).

I also replaced the call to 'wc -l' with an 'awk' command that does the same.
The prompt functions already call awk many times, so one more won't make that
much difference, plus I can better format the count anyway and ditch the shell
string manipulation.