mount and umount a remote server filesystem using SSHFS
With this (and FUSE SSHFS) I can mount any subtree of any remote host that
provides an SSH service.  The files appear mounted underneath $SSHFS_MOUNT_POINT
as-if they were local files, meaning I can edit them with a local editor or
viewer.

This is similar to Emacs' TRAMP but for all programs on the local machine.