Link LIB files under lib/
This is a fairly large change that combines three things:

 * adding a Dotfiles Step to link Libs, similar to linking Configs
 * refactoring Python init/source to put VEnvs under ~/lib/python
 * Moving the vagrant library files from link/.vagrant.d to lib/vagrant

I couldn't see a way to cleanly separate all three because they relate.

There's a new global environment variable, LIB which points to the library
directory ~/lib and is used by the dotfiles Libs step, as well as by the Python
functions.

Issue #5