fudge: deselect Init scripts that don't have ".sh" in their name
Since the Init step `source`s scripts, they must be shell scripts.  The check
for ".sh" is to automatically skip any that I've moved to a different name
because they are broken -- e.g. 50_node.bug

I'd like to exclude it from the list entirely, but that's not doable because
init_files() is called from do_stuff() which (necessarily) selects _all_ files
for a particular Step.