convert lssrc into a shell function using awk; tighten up _src()
Whenever you see a pipeline involving grep and sed and/or cut, you should know
that awk can do the same with fewer process forks, and nicer code.

It had to be a function though, because you can't make an alias that contains
both " and ' characters in it.  Besides, the GNU Bash Info manual recommends:

   For almost every purpose, shell functions are preferred over aliases.
     -- bash.info; node 6.6, Aliases

The _src() bash completion function for src() was using similar pipeline to
generate the COMPWORDS. Replaced that with a call to lssrc() (DRY) and also
removed redundant local vars