move meta-functions into source/10_meta.sh
These meta-functions are used throughout My additions to Dotfiles for
documenting and exploring bash functions from the command line. They are also
orthogonal to the rest of the Dotfiles Copy/Init/Link etc. bootstrapping logic,
so they can be moved out.

I've done this in order to better facilitate sharing individual bash modules
with friends and colleages without them needing to inherrit the whole Dotfiles
architecture if they don't want it, and are only interested in my Python module,
or the SSH stuff, for instance.

I would give them 10_meta.sh for the utility functions, along with whichever
module(s) they are actually interested in, and tell them to source both, in
alphanumeric order with a code similar to the src() function in link/.bashrc