note functions needed by bootstrap
When Dotfiles is first bootstrapped onto a new computer, the only file that
exists to begin with is bin/dotfiles itself.  It then proceeds to check thingns
out and git-clone the rest of the repository.

Therefore there is a core of shell functions at the top of the code which are
used by the rest of bin/dotfiles to bootstrap itself, and these cannot be
removed into a source/* module.  These are the OS detection routines for the
most part.

is_exe is an interesting one:  it is *not* referenced directly within
bin/dotfiless, but it *is* used in init/50_python.sh and /potentially/ in other
init modules yet to be written for other OSes.  So either it has to stay, or I
have to change the bootstrap mechanics to source at it before doing the Init
Step.

It's far simpler just to leave it here.