make ubuntu and node init steps work for Raspbian
There are some small differences between Ubuntu and Raspbian, but
they are not enough to be worried over.  So I've made the checks
on the `20_ubuntu_apt.sh` allow for Raspbian also.

Similarly for Node, the executable on Ubuntu/Raspbian is called
`nodejs` rather than `node`, otherwise it's the same

Issue #10