use pipsi for python tools in global environment
On macOS, Homebrew has removed pip from global installs.  Rather than try and
work out why or how to fix, I've used this as a prompt to switch to using pipsi
for python shell tools.  So now the python Init step involves installing pipsi
if required, and then using it to install pip, virtualenv, ipython and so on as
pipsi packages.