fix pip in Anaconda mode
PIP_REQUIRE_VIRTUALENV=true will force pip to check a venv is active, which is
going to fail when a conda env is active, even though the use of pip in this
case is fine.  So in Anaconda mode, turn this variable off, but put it back on
when leaving Anaconda.