freezenv - save a venv's installed packages (requirements) for later retrieval
Whenever Homebrew (macOS) updates python versions and libraries, it can break
your installed venvs because of broken links to old libraries.  This function is
a work-around: use it to keep the python package lists for your venvs together
with the venv, so that you can regenerate the env in a fresh one if you need to.

Supports anaconda too, though that shouldn't be necessary.